feat(post): recreated the post.
BREAKING CHANGE: Section now is deprecated
