+++ src/components/pages/feed/Post/index.ts
@@ -1,0 +1,2 @@
+export { default } from "./post"
+export type { PostProps } from "./post.types"
