+++ src/components/pages/feed/Post/post.types.ts
@@ -1,0 +1,10 @@
+import { Image, Meta } from "@/shared/data.types";
+
+export type PostProps = {
+    meta: Meta
+    image: Image;
+    title: string;
+    subtitle: string;
+
+    publishedDate: Date;
+}
