--- src/components/ui/Post/index.ts
@@ -1,2 +1,0 @@
-export { default } from "./Post"
-export type { PostProps } from "./types/PostProps"
