--- src/components/ui/Post/types/PostProps.ts
@@ -1,15 +1,0 @@
-export type PostProps = {
-    slug: string;
-    image: {
-        src: string;
-        sizes: string[];
-        alt: string;
-    }
-    title: string;
-    subtitle: string;
-    badge: {
-        label: string;
-        color: string;
-    };
-    publishedDate: string;
-}
