+++ src/configs/url.config.ts
@@ -1,0 +1,6 @@
+import { Heading } from "@/shared/data.types";
+
+export function getHeadingUrl(heading: Heading, url: string): string {
+    return `/${heading}/${url}`
+}
+
