+++ src/shared/data.types.ts
@@ -1,0 +1,11 @@
+export type Image = {
+    src: string;
+    alt: string;
+}
+
+export type Heading = "amazes" | "informs" | "amuses" | "inspires"
+
+export type Meta = {
+    heading: Heading;
+    slug: string;
+}
