+++ src/types/Section.ts
@@ -1,1 +1,5 @@
+/**
+ * @deprecated use shared/data.types.ts (Heading)
+ *
+ */
 export type Section = "Amazes" | "Informs" | "Amuses" | "Inspires" | (string & {})
